+++ src/inline_requests/__init__.py
@@ -11,7 +11,7 @@ from .utils import get_args
 
 __author__ = 'Rolando Espinoza'
 __email__ = 'rolando at rmax.io'
-__version__ = '0.3.1-dev'
+__version__ = '0.3.1'
 
 __all__ = ['inline_requests']